backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
SQLAlchemy==2.0.25
psycopg[binary]==3.1.16
